# llm-service/docs/contracts/llm-service-v1-moderacion.openapi.yaml
openapi: 3.0.3
info:
  title: llm-service — Moderación de chat (EP-08)
  version: 1.1.0
  description: |
    Contrato del API de moderación implementado en `llm-service` (EP-08). Reemplaza al borrador
    `llm-service-v1-moderacion-borrador.yaml` (que describía `/ai/moderador`, ya inexistente).
    Los ejemplos son respuestas capturadas del servicio real (salvo `pendingTimeout`, ilustrativo).

    **Flujo:** `chat-service` pide una decisión síncrona (`POST /decisions`) ANTES de entregar el
    mensaje. Si es `BLOCK` se abre un incidente; el alumno puede apelar (`POST /appeals`) y el
    docente del curso lo revisa y resuelve (`POST /incidents/{id}/resolve`).
    `PENDING` y `PENDING_REVIEW` = el mensaje NO se publica hasta que un docente lo revise.

    **Cambio v1.1.0 (breaking, acordado con chat):** `sender_id` es obligatorio en
    `POST /decisions`. Sin él no se puede registrar de quién es el incidente ni validar que solo
    su dueño apele (anti-IDOR).

    **Autenticación:** este servicio NO valida JWT; confía en los headers `X-*` que inyecta el
    API Gateway (DEC-08). Para probar directo contra el servicio usar esos headers (ver `securitySchemes`).

    **Rutas:** todas también existen bajo el prefijo del Gateway `/api/llm` (ej. `/api/llm/moderation/v1/decisions`).
servers:
  - url: http://localhost:8086
    description: llm-service directo (overlay compose.debug.yaml)
  - url: http://localhost:4011
    description: Mock Prism (ver docs/contracts/MOCK.md)
...
security:
  - GatewayHeaders: []
paths:
  /moderation/v1/decisions:
    post:
      tags: [Decisiones]
      operationId: decidirModeracion
      summary: Decide ALLOW / BLOCK / PENDING antes de entregar un mensaje (chat-service)
      description: |
        Requiere principal de **servicio** con scope `moderation:decide`. Idempotente por `message_id`:
        reenviarlo devuelve la decisión original aunque cambie el texto (la discrepancia se alerta en logs).
        Timeout interno 800 ms: si vence responde `PENDING` y el chat NO debe publicar. Nunca se aprueba
        automáticamente un mensaje que no pudo verificarse.
      parameters:
        - $ref: '#/components/parameters/PrincipalType'
        - $ref: '#/components/parameters/ServiceId'
        - $ref: '#/components/parameters/ServiceScopes'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ModerationDecisionRequest' }
            examples:
              limpio:
                value:
                  message_id: msg-001
                  course_id: curso-42
                  sender_id: alumno-1
                  sender_role: student
                  text: Hola profe, cuando es la entrega?
              spam:
                value:
                  message_id: msg-002
                  course_id: curso-42
                  sender_id: alumno-1
                  sender_role: student
                  text: mira http://a.com http://b.com http://c.com ya
      responses:
        '200':
          description: Decisión emitida (o recuperada por idempotencia)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ModerationDecisionResponse' }
              examples:
                allow:
                  value: { message_id: msg-001, decision: ALLOW, reason_code: CLEAN, classifier_used: deterministic, latency_ms: 3, incident_id: null }
                block:
                  value: { message_id: msg-002, decision: BLOCK, reason_code: SPAM, classifier_used: deterministic, latency_ms: 1, incident_id: eee5cf4d-2903-42dd-bcc8-fead34b06450 }
                pendingReviewDegradado:
                  summary: Clasificador contextual caído (modo degradado)
                  value: { message_id: msg-009, decision: PENDING_REVIEW, reason_code: NEEDS_REVIEW, classifier_used: fallback, latency_ms: 599, incident_id: d2097bef-cd56-425a-a3ff-415b00bbfff9, degradation_reason: CONTEXTUAL_UNAVAILABLE }
                pendingTimeout:
                  value: { message_id: msg-010, decision: PENDING, reason_code: ENGINE_UNAVAILABLE, classifier_used: fallback, latency_ms: 801, incident_id: 7d6f0a9e-1b2c-4d3e-8f40-5a6b7c8d9e01, degradation_reason: FULL_ENGINE_UNAVAILABLE }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    GatewayHeaders:
      type: apiKey
      in: header
      name: X-Principal-Type
      description: |
        Headers inyectados por el API Gateway. Servicio: `X-Principal-Type: service`, `X-Service-Id`,
        `X-Service-Scopes: moderation:decide`. Usuario: `X-Principal-Type: user`, `X-User-Id`, `X-User-Roles`
        (docente además `X-Teacher-Course-Ids: curso-42,curso-43`). Sin headers: 401.
  parameters:
    PrincipalType: { name: X-Principal-Type, in: header, required: true, schema: { type: string, enum: [service, user] } }
    ServiceId: { name: X-Service-Id, in: header, required: true, schema: { type: string, example: chat-service } }
    ServiceScopes: { name: X-Service-Scopes, in: header, required: true, schema: { type: string, example: 'moderation:decide' } }
...
  responses:
    BadRequest:
      description: Payload inválido
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Problem' }
          example: { type: about:blank, title: Bad Request, status: 400, detail: Los datos enviados no son válidos o contienen campos obligatorios ausentes., error: validation_error }
    Unauthorized:
      description: Sin identidad o sin el scope requerido
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Problem' }
          examples:
            sinHeaders: { value: { type: 'https://tpi.utn.frc/errors/no-autenticado', title: No autenticado, status: 401, detail: El request no trae headers de identidad validados. } }
            scopeFaltante: { value: { type: about:blank, title: Unauthorized, status: 401, detail: 'Missing scope: moderation:decide' } }
...
  schemas:
    ModerationDecisionRequest:
      type: object
      required: [message_id, course_id, sender_id, sender_role, text]
      properties:
        message_id: { type: string, description: Clave de idempotencia }
        course_id: { type: string }
        sender_id: { type: string, maxLength: 255, description: 'Id del usuario que escribió el mensaje. NUEVO v1.1.0; dueño del incidente.' }
        sender_role: { type: string, example: student, description: 'student | teacher | system' }
        text: { type: string, minLength: 1, maxLength: 4096 }
        context_flags: { type: object, additionalProperties: true, description: Solo contexto inmediato, nunca historial }
    ModerationDecisionResponse:
      type: object
      required: [message_id, decision, reason_code, classifier_used, latency_ms, incident_id]
      properties:
        message_id: { type: string }
        decision: { type: string, enum: [ALLOW, BLOCK, PENDING, PENDING_REVIEW] }
        reason_code: { type: string, example: SPAM, description: 'CLEAN, SPAM, OFFENSIVE, CODE_OBFUSCATION, NEEDS_REVIEW, ENGINE_UNAVAILABLE…' }
        classifier_used: { type: string, enum: [deterministic, contextual, fallback] }
        latency_ms: { type: integer, format: int64 }
        incident_id: { type: string, format: uuid, nullable: true, description: 'Presente si decision != ALLOW' }
        degradation_reason: { type: string, description: 'Solo en modo degradado: CONTEXTUAL_UNAVAILABLE | FULL_ENGINE_UNAVAILABLE' }
...
    Problem:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        requestId: { type: string }
        error: { type: string }